Refactor neighbors
