+++ skfda/ml/regression/__init__.py
@@ -1,9 +1,8 @@
 
 
-from .._neighbors import (KNeighborsScalarRegressor,
-                          RadiusNeighborsScalarRegressor,
-                          KNeighborsFunctionalRegressor,
-                          RadiusNeighborsFunctionalRegressor)
+from ..._neighbors import (KNeighborsScalarRegressor,
+                           RadiusNeighborsScalarRegressor,
+                           KNeighborsFunctionalRegressor,
+                           RadiusNeighborsFunctionalRegressor)
 
 from .linear_model import LinearScalarRegression
-
